refactory : enum Type 관련 dto 계층 depth 추가
