+++ src/main/java/sopt/org/homepage/project/dto/GetProjectsRequestDto.java
@@ -4,6 +4,8 @@ import io.swagger.v3.oas.annotations.Parameter;
 import lombok.Getter;
 import org.springframework.validation.annotation.Validated;
 import sopt.org.homepage.common.dto.PaginateRequest;
+import sopt.org.homepage.project.dto.type.ProjectType;
+import sopt.org.homepage.project.dto.type.ServiceType;
 
 @Validated
 @Getter
