+++ src/main/java/sopt/org/homepage/project/dto/ProjectDetailResponse.java
@@ -6,6 +6,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
+import sopt.org.homepage.project.dto.type.ServiceType;
 
 @ToString(callSuper = true)
 @EqualsAndHashCode(callSuper = true)
